Do some renaming. Move QiskitService components back.
