+++ org.quantil.qprov.collector/src/main/java/org/quantil/qprov/collector/providers/ibmq/service/QiskitServiceResult.java
@@ -17,18 +17,23 @@
  * limitations under the License.
  *******************************************************************************/
 
-package org.quantil.qprov.collector.providers.qiskit.service;
+package org.quantil.qprov.collector.providers.ibmq.service;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Data
+@NoArgsConstructor
 @AllArgsConstructor
-public class QiskitServiceRequest {
+public class QiskitServiceResult {
 
-    @JsonProperty("qpu-name")
-    private String qpu_name;
+    private String id;
 
-    private String token;
+    private boolean complete;
+
+    private Map<String, Object> result = new HashMap<>();
 }
